requirements.txt
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
colorama==0.4.6
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
textdistance==4.6.2
requests==2.32.3